@@ -18,6 +18,9 @@ public class Film {
     private String description;
     private LocalDate releaseDate;
     private Integer duration;
+    private MpaaRate mpa;
+
+    private List<Genre> genres;
 
     public void addLike(int userId) {
         userLikes.add(userId);
@@ -27,7 +30,19 @@ public class Film {
         userLikes.remove(userId);
     }
 
-    public int getLikesCount(){
+    public int getLikesCount() {
         return userLikes.size();
     }
+
+    public Map<String, Object> toMap() {
+        Map<String, Object> values = new HashMap<>();
+        values.put("name", name);
+        values.put("description", description);
+        values.put("release_date", releaseDate.toString());
+        values.put("duration", duration);
+        values.put("mpaa_rate_id", mpa.getId());
+        return values;
+    }
+
+
 }
