@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.user.User;
@@ -19,7 +20,7 @@ public class UserService {
     private final UserStorage storage;
 
     @Autowired
-    public UserService(UserStorage storage) {
+    public UserService(@Qualifier("userDbStorage") UserStorage storage) {
         this.storage = storage;
     }
 
@@ -47,40 +48,34 @@ public class UserService {
             log.warn("Film ID below 0");
             throw new ValidationException("Film id should be positive", "id");
         }
-        User user = storage.findUserById(userId);
+        User user = storage.getUserById(userId);
         storage.remove(user);
         return UserDTO.fromUser(user);
     }
 
     public UserDTO findUserById(int userId) {
-        return UserDTO.fromUser(storage.findUserById(userId));
+        return UserDTO.fromUser(storage.getUserById(userId));
     }
 
     public Collection<UserDTO> addFriend(int userId, int friendId) {
-        User user = storage.findUserById(userId);
-        User friend = storage.findUserById(friendId);
-        user.addFriend(friend);
-        friend.addFriend(user);
+        storage.addUserFriend(userId, friendId);
         return getFriends(userId);
     }
 
     public Collection<UserDTO> removeFriend(int userId, int friendId) {
-        User user = storage.findUserById(userId);
-        User friend = storage.findUserById(friendId);
-        user.removeFriend(friend);
-        friend.removeFriend(user);
+        storage.removeUserFriend(userId, friendId);
         return getFriends(userId);
     }
 
     public Collection<UserDTO> getFriends(int userId) {
-        return storage.findUserById(userId).getFriends().values().stream()
+        return storage.getUserFriends(userId).stream()
                 .map(UserDTO::fromUser)
                 .collect(Collectors.toList());
     }
 
     public Collection<UserDTO> getCommonFriends(int user1Id, int user2Id) {
-        Collection<User> friends1 = storage.findUserById(user1Id).getFriends().values();
-        Collection<User> friends2 = storage.findUserById(user2Id).getFriends().values();
+        Collection<User> friends1 = storage.getUserFriends(user1Id);
+        Collection<User> friends2 = storage.getUserFriends(user2Id);
         return friends1.stream().filter(friends2::contains).map(UserDTO::fromUser).collect(Collectors.toList());
     }
 
